feat: implement projectile pool manager
